+++ src/main/php/Statement.php
@@ -22,12 +22,12 @@ interface Statement
      *
      * @param   int|string  $param     the order number of the parameter or its name
      * @param   mixed       $variable  the variable to bind to the parameter
-     * @param   int|string  $type      optional  type of the parameter
+     * @param   int         $type      optional  type of the parameter
      * @param   int         $length    optional  length of the data type
      * @return  bool        true on success, false on failure
      * @throws  \stubbles\db\DatabaseException
      */
-    public function bindParam($param, &$variable, $type = null, int $length = null): bool;
+    public function bindParam($param, &$variable, int $type = null, int $length = null): bool;
 
     /**
      * bind a value to the parameter of a prepared query
@@ -37,11 +37,11 @@ interface Statement
      *
      * @param   int|string  $param  the order number of the parameter or its name
      * @param   mixed       $value  the value to bind
-     * @param   int|string  $type   optional  type of the parameter
+     * @param   int         $type   optional  type of the parameter
      * @return  bool        true on success, false on failure
      * @throws  \stubbles\db\DatabaseException
      */
-    public function bindValue($param, $value, $type = null): bool;
+    public function bindValue($param, $value, int $type = null): bool;
 
     /**
      * executes a prepared statement